Add support for prefixing bug cards in Trello with [Bug $bugnum]
This allows us to also be more flexible in searching for dupes as we can
search on the prefix (this is useful in case bug names have changed).
This preserves backwards compatibility with the previous, non-prefixed
way of doing things.

This behavior is configurable, see bugello.ini.example diff for details.

Resolves issues #32